winmm: Disable forceful termination of timer threads.
This is absolutely unnecessary and may cause resource leaks
eventually leading to the application hang or crash.

The above symptoms are known to happen in Adobe Flash applications
which use a lot of timers which are started/stopped at a very high rate.
This fix gets rid of these symptoms.