Added .gitattributes file to force CRLF on Rexx Scripts
Because git-svn is used at the dev-side, most svn properties were lost
and files were 'normalized' to LF when the repo was git-svn cloned.
This is not too big a problem, but it does impact line endings.

On OS/2, Rexx barks when scripts do not use CRLF line endings.
So, to enforce CRLF on these scripts, we add the .gitattributes file
and add the .cmd extension to it.

A native svn checkout under OS/2 also reveiled that most KMK Makefiles
were committed with CRLF, so we'll also add the .kmk extension to
.gitattributes for consistency, eventhough KMK couldn't care less.

We'll probably do a few more test-commits to see if using git-svn
does not introduce any anomalities.