Probably fixed duplicate symbols with OLE ID's
On Win32 these symbols are generated in the '.data' section of the
resulting COFF32 object-module. Linking them does not introduce
duplicate symbol definitions.

On OS/2 they are generated in the '.text' section of the resulting a.out
object-module. This causes the 'weak pre-linker' to abort on duplicate
symbols. One way to work-around this is to use the 'wl.exe' linker
directly, after conversion to OMF, which treats them as warnings.

The solution in this fix hides the symbols by reducing their visibility
to module-scope. It has not yet been determined if this fix is adequate.