usp10: Added Wine Sources (containing stubs)
The first goal is to have 'usp10' compile and build for Odin32.
The second goal is to enable the JNI-linkage for 'swt.dll'.
Instead of complicating things by starting with the latest and greatest
versions from the Wine Repo, we start with the version where Wine
started implementing 'usp10'.

Then we'll use some 'successive approximation' to find a commit in the
Wine Repo that is the best fit for Odin32 and SWT at this time. That
will be hour 'hook' into the Wine Repo for 'usp10'. Then we'll first
port the functions that SWT needs, leaving the others for a later time.

This revision builds and 'swt.dll' can resolve the 4 exported funtions.
(Testing with Snippet316)

For all the Wine Odin is going to drink we use the Wine Git Repository.
Below are the references to the GitGub mirror and the Master repo for
the sources used for this 'glass of wine'.

Wine Mirror (GitHub) Source and Commit references:
https://github.com/wine-mirror/wine/tree/54508df52715eaeb6414f9f474691fde548d8e15/dlls/usp10
https://github.com/wine-mirror/wine/commit/54508df52715eaeb6414f9f474691fde548d8e15

Wine Master (GitWeb) Source and Commit references:
http://source.winehq.org/git/wine.git/tree/54508df52715eaeb6414f9f474691fde548d8e15:/dlls/usp10
http://source.winehq.org/git/wine.git/commit/54508df52715eaeb6414f9f474691fde548d8e15

Note:
A header-file for 'usp10' was already committed here:
http://trac.netlabs.org/odin32/changeset?reponame=&old=22076%40%2F&new=22076%40%2F
for getting 'swt.dll' to build. It will be synced from the Wine Repo
when we have found our 'hook' commit.