Work around Open Watcom v1.9 (owcc) bug                        [apitest]
The 'owcc' compiler front-end of v1.9 does not handle '/' in source
specifications correctly. So the 'subst' macro is used to revert to '\'
for those cases. This seems to have been fixed in v2.0-beta.

Also make Odin Based and Win32 variants use the built Odin instead of
the stock release.

And added a target to start Win32 variants using 'pe'.